added unit tests to test Anagrams
