+++ package.json
@@ -4,8 +4,9 @@
   "description": "JavaScript Interview Series with Test Driven Design (TDD)",
   "main": "app.js",
   "scripts": {
-    "start": "node app",
-    "test": "jest -watch"
+    "start": "nodemon app",
+    "test": "jest --watch",
+    "lint": "eslint ."
   },
   "repository": {
     "type": "git",
@@ -32,8 +33,37 @@
   },
   "homepage": "https://github.com/rlama7/JavaScriptInterviewSeries#readme",
   "devDependencies": {
+    "eslint": "^7.19.0",
+    "eslint-config-prettier": "^7.2.0",
+    "eslint-plugin-prettier": "^3.3.1",
     "jest": "^26.6.3",
     "nodemon": "^2.0.7",
+    "prettier": "^2.2.1",
     "supertest": "^6.1.3"
+  },
+  "eslintConfig": {
+    "parserOptions": {
+      "ecmaVersion": 6
+    },
+    "extends": [
+      "eslint: recommended",
+      "plugin: prettier/recommended"
+    ],
+    "plugins": [
+      "prettier"
+    ],
+    "env": {
+      "node": true,
+      "es6": true,
+      "jest": true
+    },
+    "rules": {
+      "eqeqeq": "warn",
+      "prettier/prettier": "warn"
+    },
+    "prettier": {
+      "singleQuote": true,
+      "printWidth": 120
+    }
   }
 }
